+++ R/reexports.R
@@ -13,3 +13,7 @@ stats::Gamma
 #' @importFrom brms bf
 #' @export
 brms::bf
+
+#' @importFrom distspec as_dist_spec
+#' @export
+distspec::as_dist_spec