+++ inst/make_hexsticker.R
@@ -11,8 +11,10 @@ secondary_dist <- data.frame(mu = 1.8, sigma = 0.5) |>
 
 obs <- outbreak |>
   simulate_secondary(
-    meanlog = secondary_dist$mu[[1]],
-    sdlog = secondary_dist$sigma[[1]]
+    distspec::LogNormal(
+      meanlog = secondary_dist$mu[[1]],
+      sdlog = secondary_dist$sigma[[1]]
+    )
   ) |>
   mutate(
     ptime_lwr = floor(.data$ptime),